Massive overhaul of Black Hole Suite
Bandwidth Monitoring processing has been removed from the Main Form
(with one exception) and the project has been reorganized into three
parts:
1. BlackHoleClient:
- Displays the bandwidth statistics and usage.
- All processing has been removed from this forum, minus Current
Upload/Download Speeds which is optional, and disabled by default.
- Some features like Clearing Bandwidth History are still WIP.

2. BlackHoleService:
- A newly created windows serviced designed to handle the vast majority
of the processing and logging with little impact on performance.

3. BlackHoleLib:
- This is the library that ties the project together. All important
classes for the forms and service are located here.